fix: problem with set Default billing Address
